now database deleter returns false if track wasn't found

+++ Mp3Player/DataBase/DataBaseDeleter.cs
@@ -12,9 +12,11 @@ public class DataBaseDeleter: IDataBaseDeleter {
         _path = path;
     }
 
-    public async Task DeleteTrack(string id) {
+    public async Task<bool> DeleteTrack(string id) {
         var files = Directory.GetFiles(_path, $"{id}.json", SearchOption.AllDirectories);
+        if (files.Length == 0) return false; 
         var deleteTasks = files.Select(file => Task.Run(() => File.Delete(file)));
         await Task.WhenAll(deleteTasks);
+        return true;
     }
 }
